updated artist application form
